type-improvements, remove excess functions
remove `hitTest.ts`, `createTransformedPath`
instead of keeping a memoized version of a transformed path
with `createTransformedPath`, we now transform the canvas-context
while rendering and doing the hitTest.

I removed `hitTest` in favor of inlining it in `createPath2D`
because it was not conform to the '1 single responsibility' rule
and it is only used once in the codebase.